fix issue with "switchable" trie implementations
this commit fixes a flaw in the method for enabling
an IAL user to switch between a "fast" version of trie
and the regular version. previously, the trie-fast
module would load the trie module, hide the
lookup/insert/remove functions, and override them with its
own versions. however this didn't cause internal functions
within the trie module to use the alternative implementations,
and so the switch was only partial.

this issue is now resolved by factoring trie into a "core"
module with the basic data structure and basic functions,
and a "functions" module that has more complex trie functions.
the "functions" module is parameterized over the
lookup/insert/remove definitions, and so the trie/trie-fast
modules can provide their own implementations.